Add `rw` command
Users should be reminded to issue `rw` in case the first thing they want to do is change password.
It is mentioned in the login screen, and explicit for step 5 , but not mentioned here. 
The error message gives if the root filesystem is read-only is  "passwd: Authentication token lock busy" is particularly obscure.