Add exfat warning for USB thumb drives (#1444)
By default, exfat drives are mounted owned by root, with 755 perms
on the mountpoint.  Rather than going into setting 'uid' and 'gid'
params, which rely on the uid and gid of the kvm user never changing,
I thought it would be simpler to just give users the hint to set the
default umask to 0000 (aka, 777) for all files. This allows the
thumb drive to be visible as read-write.

Signed-off-by: Rob Thomas <xrobau@gmail.com>